Created a branch for Angular front-end project
These are the updated files for our Angular front-end. Follow the README file for Angular for a guide on setting up this project locally. The "styles.css" and "website-background-img" files should go directly in your "src" folder once you create the angular application. The rest of the files go in "fish-tank-project/src/app" folder.