Added user input fields and buttons
Styled user input fields so "email" sits in a row by itself, and "first name" sits on a row with "last name." Added two buttons at the bottom of the form, "submit" and "clear." Added functionality to these buttons via the app.component.ts file. Follow the steps in the Angular README file in the master branch to set up an Angular project. Then, replace the files in your "app" folder with these to get the latest version of the website.