URL Column and Grid Link
I took the URL column out of the table because I did not like how it looked, but, if we want it back, I can put the column back in pretty easily. Also, when an internship is clicked on (no matter where the user clicks within the row), a new internet window/tab will open to the information page on the FIS website.