Added embedded video support in webpage
Added HTML5 tag to support an embedded video file. If you are using chrome the video will have to be hosted online, as Chrome does not allow you to display local media in-browser as it is a security risk. If you have a browser that allows you to display local media, simply provide the file path to the video file to view the video in the webpage.